refactor(shared): 컨텐츠 형식의 도메인 상태 enum 공통 으로 변경
