+++ src/main/java/com/backend/immilog/shared/enums/ContentStatus.java
@@ -1,0 +1,8 @@
+package com.backend.immilog.shared.enums;
+
+public enum ContentStatus {
+    NORMAL,
+    MODIFIED,
+    REPORTED,
+    DELETED
+}
